+++ jpsonic-main/src/main/java/org/airsonic/player/service/search/SearchServiceImpl.java
@@ -250,8 +250,7 @@ public class SearchServiceImpl implements SearchService {
     public List<MediaFile> getRandomSongs(int count, int offset, int casheMax, List<MusicFolder> musicFolders) {
 
         final List<MediaFile> result = new ArrayList<>();
-        Consumer<List<Integer>> addSubToResult = (ids) -> ids
-                .subList((int) offset, Math.min(ids.size(), (int) (offset + count)))
+        Consumer<List<Integer>> addSubToResult = (ids) -> ids.subList(offset, Math.min(ids.size(), offset + count))
                 .forEach(id -> util.addIgnoreNull(result, IndexType.SONG, id));
         util.getCache(RandomCacheKey.SONG, casheMax, musicFolders).ifPresent(addSubToResult);
         if (0 < result.size()) {
@@ -305,7 +304,7 @@ public class SearchServiceImpl implements SearchService {
 
         final List<MediaFile> result = new ArrayList<>();
         Consumer<List<MediaFile>> addSubToResult = (files) -> {
-            List<MediaFile> sub = files.subList((int) offset, min(files.size(), (int) (offset + count), casheMax));
+            List<MediaFile> sub = files.subList(offset, min(files.size(), offset + count, casheMax));
             result.addAll(sub);
         };
 
@@ -385,8 +384,7 @@ public class SearchServiceImpl implements SearchService {
     public List<Album> getRandomAlbumsId3(int count, int offset, int casheMax, List<MusicFolder> musicFolders) {
 
         final List<Album> result = new ArrayList<>();
-        Consumer<List<Integer>> addSubToResult = (ids) -> ids
-                .subList((int) offset, Math.min(ids.size(), (int) (offset + count)))
+        Consumer<List<Integer>> addSubToResult = (ids) -> ids.subList(offset, Math.min(ids.size(), offset + count))
                 .forEach(id -> util.addIgnoreNull(result, IndexType.ALBUM_ID3, id));
         util.getCache(RandomCacheKey.ALBUM, casheMax, musicFolders).ifPresent(addSubToResult);
         if (0 < result.size()) {
@@ -450,7 +448,7 @@ public class SearchServiceImpl implements SearchService {
 
         final List<MediaFile> result = new ArrayList<>();
         Consumer<List<MediaFile>> addSubToResult = (mediaFiles) -> result
-                .addAll(mediaFiles.subList((int) offset, Math.min(mediaFiles.size(), (int) (offset + count))));
+                .addAll(mediaFiles.subList(offset, Math.min(mediaFiles.size(), offset + count)));
         util.getCache(genres, musicFolders, IndexType.ALBUM).ifPresent(addSubToResult);
         if (0 < result.size()) {
             return result;
@@ -509,7 +507,7 @@ public class SearchServiceImpl implements SearchService {
 
         final List<MediaFile> result = new ArrayList<>();
         Consumer<List<MediaFile>> addSubToResult = (mediaFiles) -> result
-                .addAll(mediaFiles.subList((int) offset, Math.min(mediaFiles.size(), (int) (offset + count))));
+                .addAll(mediaFiles.subList(offset, Math.min(mediaFiles.size(), offset + count)));
         util.getCache(genres, musicFolders, IndexType.SONG).ifPresent(addSubToResult);
         if (0 < result.size()) {
             return result;
